Part 1 semi-complete
Made it so that after every race, the confidence level of the horse changes, these confidence levels are then added to a file and stored. At the start of every race, the horses in the file can be selected to be used for the track. Also made it so that lanes can be empty