GUI improvements
Added a new race panel class which will hopefully take care of the logic for the GUI race, as well as added new features to the track settings page, and also fixed a lot of bugs. Currently the start race doesn't work, but prints 3 lanes(due to that being the default when GUI is initialized).